rag_service/requirements.txt
fastapi
uvicorn[standard]
weaviate-client
openai>=1.0.0
structlog
prometheus_client